remove reset password component and refactor login form
