--- src/components/forgot-password/forgotPassword.style.js
@@ -1,23 +1,0 @@
-const style = {
-  root: {
-    maxWidth: '343px',
-    padding: '55px 90px'
-  },
-  h2: {
-    paddingBottom: 2,
-    fontSize: '32px',
-    lineHeight: '40px'
-  },
-  sentPass: {
-    mb: 1,
-    padding: '18px 49px' 
-  },
-  back: {
-    textAlign: 'center',
-    alignItems: 'center',
-    textDecoration: 'underline',
-    padding: '18px 112.5px',
-  }
-}  
-  
-export default style
