+++ src/components/google-login/google-login.style.js
@@ -1,0 +1,24 @@
+const style = {
+  linesBox: {
+    margin: '23px 0',
+    display: 'flex',
+    justifyContent: 'space-between',
+    alignItems: 'center'
+  },
+  line: {
+    display: 'flex',
+    width: '34%', 
+    height: '2px', 
+    backgroundColor: 'primary.100'
+  },
+  google: {
+    marginBottom: '16px',
+    padding: '15px 77px'   
+  },
+  underlineText: {
+    fontWeight: '500',
+    textDecoration: 'underline',
+  }
+}
+
+export default style
