+++ src/components/login-dialog/LoginDialog.style.js
@@ -17,43 +17,6 @@ const style = {
   form: { 
     maxWidth: '343px',
     padding: { sm: '50px 80px', md: '0' }
-  },
-  input: {
-    maxWidth: '343px'
-  },
-  checkboxContainer: {
-    margin: '25px 0',
-    display: 'flex',
-    justifyContent: 'space-between',
-    alignItems: 'center'
-  },
-  checkboxLabel: {
-    	'& .MuiFormControlLabel-label': {
-      typography: 'subtitle2'
-    }
-  },
-  loginButton: {
-    padding: '18px 149px'
-  },
-  linesBox: {
-    margin: '23px 0',
-    display: 'flex',
-    justifyContent: 'space-between',
-    alignItems: 'center'
-  },
-  line: {
-    display: 'flex',
-    width: '34%', 
-    height: '2px', 
-    backgroundColor: 'primary.100'
-  },
-  google: {
-    marginBottom: '16px',
-    padding: '15px 77px'   
-  },
-  underlineText: {
-    fontWeight: '500',
-    textDecoration: 'underline',
   }
 }
