+++ src/components/login-form/login-form.style.js
@@ -1,0 +1,29 @@
+const style = {
+  form: { 
+    maxWidth: '343px',
+    padding: { sm: '50px 80px', md: '0' }
+  },
+  input: {
+    maxWidth: '343px'
+  },
+  checkboxContainer: {
+    margin: '25px 0',
+    display: 'flex',
+    justifyContent: 'space-between',
+    alignItems: 'center'
+  },
+  checkboxLabel: {
+    	'& .MuiFormControlLabel-label': {
+      typography: 'subtitle2'
+    }
+  },
+  loginButton: {
+    padding: '18px 149px'
+  },
+  underlineText: {
+    fontWeight: '500',
+    textDecoration: 'underline',
+  }
+}
+
+export default style
