+++ src/constants/translations/en/login.json
@@ -16,11 +16,5 @@
         "emailValid": "Enter valid email example@mail.com",
         "passwordLength": "Password has a minimum of 6 characters adm maximum 30",
         "passwordValid": "Password has at least 1 uppercase letter, 1 lowercase letter, and 1 number with no spaces"
-    },
-    "enterEmail": "Enter the email address associated with your account and we`ll send you a link to reset your password.",
-    "sendPassword": "Send me password reset link",
-    "backToLogin": "Back to Log in",
-    "passwordReset": "Password Reset",
-    "weSentEmail": "We`ve sent you an email with a password reset link to:",
-    "emailArrive": "If the email didn't arrive, give it a few minutes and check back. If it isn’t still there, check your spam folder."
+    }
 }
